pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "calculator>=1.1.0",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
    "render>=1.0.0",
]
